Remove ifdef for setting TWO_DIGIT_EXPONENT on Windows
This change was originally added to get standard two digit exponent
behaviour on Windows and Linux. However the proprietary feature was
removed with Visual Studio 2015 and the standards conforming behaviour
was made default.

Therefore for passing tests Visual Studio 2015 or greater is required,
so updated the README to reflect the dependency on newer Visual Studio.

"Exponent formatting: The %e and %E format specifiers format a floating
point number as a decimal mantissa and exponent. The %g and %G format
specifiers also format numbers in this form in some cases. In previous
versions, the CRT would always generate strings with three-digit
exponents. For example, printf("%e\n", 1.0) would print 1.000000e+000.
This was incorrect: C requires that if the exponent is representable
using only one or two digits, then only two digits are to be printed.

In Visual Studio 2005 a global conformance switch was added:
_set_output_format. A program could call this function with the argument
_TWO_DIGIT_EXPONENT, to enable conforming exponent printing. The default
behavior has been changed to the standards-conforming exponent printing
mode." - https://msdn.microsoft.com/en-us/library/bb531344(v=vs.140).aspx